Prep v1.5.1 release: CUDA 2CXM/2CUM performance fixes
Performance-only release for the CUDA backend. Fit results are unchanged:
Gpufit_Cpufit_MRI_Parity's max |param rel diff| for 2CXM (0.00794723) and
Tissue Uptake (0.00187197) are identical to six digits before and after, and
downstream Stage-D outputs are bit-identical.

Two fixes, found by profiling 2CXM on CUDA against single-threaded Cpufit,
which the GPU was losing to by 4.8-8.7x:

- The 2CXM and Tissue Uptake model kernels accumulated their convolutions
  with a bare `0.5` literal, which promotes the whole inner-loop expression
  to double. In SASS that is 1 DMUL + 4 DFMA + 13 F2F conversions per
  iteration; consumer GPUs run FP64 at 1/64 of FP32 rate, so the kernels ran
  ~14x slower than intended while producing correct answers. The literals are
  now REAL. Both loops carry a comment saying why, since no accuracy test can
  catch this class of bug.

- cuda_calc_curve_values now honours the existing backtrack_accepted mask.
  cuda_update_parameters_trial already froze an accepted fit's parameters for
  the rest of the line search, but the model was still re-evaluated for it on
  every remaining trial. nsys put that kernel at 92% of GPU time, running
  exactly 9 evaluations per LM iteration. Deliberately not applied to
  cuda_calculate_chi_squares: it is 1.2% of GPU time and its reduction calls
  sum_up_floats, whose __syncthreads() makes a partial-block early return
  unsafe when n_fits_per_block > 1.

Net on 2CXM (3000-point curves): 7.9x / 16.5x / 21.8x faster at 24 / 240 /
2400-voxel batches, taking Gpufit from slower than single-threaded Cpufit to
4.6x faster on large batches. Tissue Uptake is 9.4x faster on 600-point
curves; its accelerated path has no dense-grid upsample, so its n_points
comes straight from the acquisition.

Verified: ctest 5/5, Gpufit_Cpufit_MRI_Parity PASS on all six MRI models
(128/128 jointly converged each), and 14 cross-backend tests in the
downstream ROCKETSHIP suite covering all five accelerated DCE models.

Bump version 1.5.0 -> 1.5.1 across all hardcoded locations and add the
README changelog entry.

+++ Gpufit/models/tissue_uptake.cuh
@@ -33,7 +33,7 @@ __device__ void calculate_tissue_uptake (               // function name
 	REAL const E = parameters[0];
 	REAL const vp = parameters[1];
 	REAL const Fp = parameters[2];
-	REAL const eps = 1e-12;
+	REAL const eps = (REAL)1e-12f;
 	REAL const oneME = 1 - E;
 
 	REAL * current_derivative = derivative + point_index;
@@ -61,9 +61,11 @@ __device__ void calculate_tissue_uptake (               // function name
 		REAL const lag_prev = T[point_index] - T[i - 1];
 		REAL const e_i = exp(-rp * lag);
 		REAL const e_p = exp(-rp * lag_prev);
-		G  += 0.5 * spacing * (Cp[i] * e_i + Cp[i - 1] * e_p);
-		Gp += 0.5 * spacing * (Cp[i] * (-lag) * e_i + Cp[i - 1] * (-lag_prev) * e_p);
-		U  += 0.5 * spacing * (Cp[i] + Cp[i - 1]);
+		// Literals must stay REAL: a bare 0.5 promotes the accumulation to double, which is
+		// correct but ~14x slower on consumer GPUs (1/64 FP64 rate). Same in 2cxm.
+		G  += (REAL)0.5f * spacing * (Cp[i] * e_i + Cp[i - 1] * e_p);
+		Gp += (REAL)0.5f * spacing * (Cp[i] * (-lag) * e_i + Cp[i - 1] * (-lag_prev) * e_p);
+		U  += (REAL)0.5f * spacing * (Cp[i] + Cp[i - 1]);
 	}
 
 	value[point_index] = E * Fp * U + Fp * oneME * G;
